Controllers: UsuariosController: Añadir limitaciones por rol
Vamos a continuar limitando el acceso desde el lado del servidor haciendo comprobaciones de variables de sesión antes de realizar alguna acción, esto con el fin de evitar que el "Personal" o los "Supervisores" realicen acciones CRUD a las cuales no están autorizados dentro de la gestión de usuarios.

Para esto comprobaremos si la variable de sesión existente tiene por rol el valor de "Administrador", si es así, el proceso del metodo se ejecutará, caso contrario, se redireccionará al usuario a otra sección.

+++ app/Http/Controllers/ReclamosController.php
@@ -148,9 +148,18 @@ class ReclamosController extends Controller
      */
     public function destroy(Reclamos $reclamosItem)
     {
-         //Realizamos la consulta Eloquent
-            Reclamos::destroy($reclamosItem->id);
-        //Redireccionamos    
-            return redirect()->back()->with('resultado', 'El reclamo ha sido eliminado correctamente'); //Devolvemos el mensaje de resultados a la vista 'reclamos'
+    //Comprobamos el rol antes de devolver la vista
+        if(session()->get('sesion')['rol'] == 'personal' || session()->get('sesion')['rol'] == 'administrador'){   
+            
+             //Realizamos la consulta Eloquent
+                Reclamos::destroy($reclamosItem->id);
+            //Redireccionamos    
+                return redirect()->back()->with('resultado', 'El reclamo ha sido eliminado correctamente'); //Devolvemos el mensaje de resultados a la vista 'reclamos'
+
+    //Redireccionamos en caso de que el rol no sea un "personal" o "administrador"   
+        }elseif(session()->get('sesion')['rol'] == 'supervisor'){
+                return redirect()->route('medidores.index');
+        }   
+
     }
 }
